feat: add external update service
