--- UpdaterHost/App.xaml
@@ -1,5 +1,0 @@
-<Application x:Class="UpdaterHost.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-</Application>
-
