--- UpdaterHost/App.xaml.cs
@@ -1,9 +1,0 @@
-using System.Windows;
-
-namespace UpdaterHost
-{
-    public partial class App : Application
-    {
-    }
-}
-
