--- UpdaterHost/ProgressWindow.xaml
@@ -1,11 +1,0 @@
-<Window x:Class="UpdaterHost.ProgressWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Atualizando" Height="150" Width="400"
-        WindowStartupLocation="CenterScreen">
-    <StackPanel Margin="20">
-        <TextBlock x:Name="StatusText" Text="Iniciando..." Margin="0,0,0,10"/>
-        <ProgressBar IsIndeterminate="True" Height="20"/>
-    </StackPanel>
-</Window>
-
