--- UpdaterHost/ProgressWindow.xaml.cs
@@ -1,18 +1,0 @@
-using System.Windows;
-
-namespace UpdaterHost
-{
-    public partial class ProgressWindow : Window
-    {
-        public ProgressWindow()
-        {
-            InitializeComponent();
-        }
-
-        public void SetStatus(string message)
-        {
-            StatusText.Text = message;
-        }
-    }
-}
-
