+++ leituraWPF.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.14.36408.4
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "leituraWPF", "leituraWPF\leituraWPF.csproj", "{87ED0F54-F906-4266-849A-B36AC214EA3A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UpdaterHost", "UpdaterHost\UpdaterHost.csproj", "{1C1743DA-FAE2-4292-B598-D85851431DF5}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{87ED0F54-F906-4266-849A-B36AC214EA3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{87ED0F54-F906-4266-849A-B36AC214EA3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{87ED0F54-F906-4266-849A-B36AC214EA3A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1C1743DA-FAE2-4292-B598-D85851431DF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C1743DA-FAE2-4292-B598-D85851431DF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C1743DA-FAE2-4292-B598-D85851431DF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C1743DA-FAE2-4292-B598-D85851431DF5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
